requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
